+++ money-transfer-rest-api/src/main/java/com/lmartino/bank/rest/dto/MakeTransferDto.java
@@ -3,12 +3,14 @@ package com.lmartino.bank.rest.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 public class MakeTransferDto {
     private String fromAccountId;
     private String toAccountId;
-    private Double amount;
+    private BigDecimal amount;
     private String description;
 
 }
